# CSHARP-1972: Further performance optimization of the Hex conversion methods in BsonUtils