May I suggest another performance optimization around the hex string to byte array (and back) conversion methods in `BsonUtils`.

This change introduces some precalculated jump tables that cost a tiny little bit of memory but give an dramatic boost in performance.

I have put together a demo solution using Benchmark.NET that proves the correctness and shows the performance improvement here: https://github.com/dnickless/BsonUtilsHexConversionPerformance