@damieng 2024-02-07
Thanks for the PR and sorry it was not looked at in a timely manner.

We will be moving to .NET 5 in an upcoming release and will use the Convert.ToHex methods that offer further performance gains.